refactor(routing): reorganize routing structure
Seperate router controllers for each module into its own file. Create a manageable twitter package which exposes router and submodules. Expose router directly from twitter package into to index.
