+++ ApiDesign/twitter/index.js
@@ -1,6 +1,4 @@
 module.exports = {
-    basics: require('./basics'),
-    tweets: require('./tweets'),
-    trends: require('./trends')
+  modules: require('./modules')
+  router: require(.'/controllers')
 }
-
