+++ ApiDesign/twitter/modules/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+    basics: require('./basics'),
+    tweets: require('./tweets'),
+    trends: require('./trends')
+}
+
